% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiabPred.R
\name{DiabPred}
\alias{DiabPred}
\title{DiabPred}
\usage{
DiabPred(genofile)
}
\arguments{
\item{genofile}{An object of class \code{SNPGDSFileClass} loaded into the environment with \code{snpgdsOpen} from the \strong{SNPRelate} package.}
}
\value{
A data frame with individuals in rows and 3 colums: 
\itemize{ 
  \item \code{prediction}: First column with the actual phenotypic output indicating wether the individual is diabetic or non-diabetic.
  \item \code{non.diabetic}: Probability of the individual being non-diabetic
  \item \code{diabetic}: Probability of the individual being diabetic.
  }
}
\description{
Prediction of type 2 diabetes based on genomic data with deep learning algorithms
}
\details{
The function described here makes use of a deep learning model trained using the h2o package to predict type 2 diabetes based on selected SNPs.
}
