% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.R
\docType{data}
\name{snps}
\alias{snps}
\title{SNP info}
\format{A data.frame consisting of 8573 rows and 6 columns

\describe{ 
  \item{chr}{Chromosome where the SNP is located}
  \item{snp_id}{SNP reference ID}
  \item{chr_start}{Start position of the SNP in the chromosome}
  \item{chr_end}{End position of the SNP in the chromosome}
  \item{ref}{Reference allele of the SNP}
  \item{alt}{Alternative allele of the SNP}
  }}
\source{
\url{http://www.type2diabetesgenetics.org/}
}
\usage{
snps
}
\description{
This dataset consists on the SNPs that were used to build the deep learning model used for prediction.
}
\details{
List of SNPs used to build this package's model
}
\keyword{datasets}
